Swap utf-8 progbar character for ASCII
In terminal work. printing Unicode characters can be problematic. In particular, using a Unicode character for the progressbar causes encoding errors if the environment is not set up properly:

```
UnicodeEncodeError: 'ascii' codec can't encode character '\u2588' in position 15: ordinal not in range(128)
```

This uses a typical ASCII character for the progress bar so that we have less to worry about when running Edward.